- ported to 4.7; old-style Makefile.ams & friends preserved or moved to Makefile.am-4.6

+++ network/yf_rtemscompat.h
@@ -6,9 +6,9 @@
 
 typedef unsigned char	u8;
 typedef unsigned short	u16;
-typedef unsigned long	u32;
+typedef unsigned 	u32;
 typedef char		s8;
 typedef short		s16;
-typedef long		s32;
+typedef int			s32;
 
 #endif
